pyproject.toml
[project]
name = "typofixer"
version = "0.1.0"
description = "A simple tool to fix typos w=using LLM and seeing the difference."
authors = [{ name = "ddorn", email = "diego.dorn@free.fr" }]
requires-python = "~=3.11"
readme = "README.md"
license = "WTFPL"
dependencies = [
    "openai>=1.41.1,<2",
    "anthropic>=0.34.1,<0.35",
    "streamlit>=1.37.1,<2",
    "tiktoken>=0.6.0,<0.7",
    "python-dotenv>=1.0.1",
]

[project.scripts]
typofixer = "typofixer.cli:cli"

[dependency-groups]
dev = [
    "ipykernel>=6.29.4,<7",
    "pre-commit>=3.7.0,<4",
]

[tool.hatch.build.targets.wheel]
packages = ["typofixer"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
